Free up the `Saml` module by extracting `XML` into the `SamlIdp` namespace.
